Moved configure_computer to the Computer orm class
This seems like a more sensible place for it
