+++ aiida/backends/tests/computer.py
@@ -57,3 +57,51 @@ class TestComputer(AiidaTestCase):
 
         with self.assertRaises(NotExistent):
             self.backend.computers.get(comp_pk)
+
+
+class TestComputerConfigure(AiidaTestCase):
+
+    def setUp(self):
+        """Prepare current user and computer builder with common properties."""
+        from aiida.control.computer import ComputerBuilder
+
+        backend = self.backend
+        self.comp_builder = ComputerBuilder(label='test', description='Test Computer', enabled=True, hostname='localhost')
+        self.comp_builder.scheduler = 'direct'
+        self.comp_builder.work_dir = '/tmp/aiida'
+        self.comp_builder.prepend_text = ''
+        self.comp_builder.append_text = ''
+        self.comp_builder.mpiprocs_per_machine = 8
+        self.comp_builder.mpirun_command = 'mpirun'
+        self.comp_builder.shebang = '#!xonsh'
+        self.user = backend.users.get_automatic_user()
+
+    def test_configure_local(self):
+        """Configure a computer for local transport and check it is configured."""
+        self.comp_builder.label = 'test_configure_local'
+        self.comp_builder.transport = 'local'
+        comp = self.comp_builder.new()
+        comp.store()
+
+        comp.configure()
+        self.assertTrue(comp.is_user_configured(self.user))
+
+    def test_configure_ssh(self):
+        """Configure a computer for ssh transport and check it is configured."""
+        self.comp_builder.label = 'test_configure_ssh'
+        self.comp_builder.transport = 'ssh'
+        comp = self.comp_builder.new()
+        comp.store()
+
+        comp.configure(username='radames', port='22')
+        self.assertTrue(comp.is_user_configured(self.user))
+
+    def test_configure_ssh_invalid(self):
+        """Try to configure computer with invalid auth params and check it fails."""
+        self.comp_builder.label = 'test_configure_ssh_invalid'
+        self.comp_builder.transport = 'ssh'
+        comp = self.comp_builder.new()
+        comp.store()
+
+        with self.assertRaises(ValueError):
+            comp.configure(username='radames', invalid_auth_param='TEST')
