+++ aiida/control/computer.py
@@ -10,53 +10,7 @@
 """Manage computer objects with lazy loading of the db env"""
 from __future__ import absolute_import
 from aiida.cmdline.utils.decorators import with_dbenv
-from aiida.common.exceptions import NotExistent
 from aiida.utils.error_accumulator import ErrorAccumulator
-
-
-def configure_computer(computer, user=None, **kwargs):
-    """Configure a computer for a user with valid auth params passed via kwargs."""
-    from aiida.orm.backend import construct_backend
-
-    transport_cls = computer.get_transport_class()
-    backend = construct_backend()
-    user = user or backend.users.get_automatic_user()
-
-    try:
-        authinfo = computer.get_authinfo(user)
-    except NotExistent:
-        authinfo = backend.authinfos.create(computer, user)
-
-    auth_params = authinfo.get_auth_params()
-    valid_keys = set(transport_cls.get_valid_auth_params())
-
-    if not set(kwargs.keys()).issubset(valid_keys):
-        invalid_keys = [key for key in kwargs if key not in valid_keys]
-        raise ValueError('{transport}: recieved invalid authentication parameter(s) "{invalid}"'.format(
-            transport=transport_cls, invalid=invalid_keys))
-
-    if valid_keys:
-        auth_params.update(kwargs)
-        authinfo.set_auth_params(auth_params)
-        authinfo.store()
-
-    return authinfo
-
-
-def get_computer_configuration(computer, user=None):
-    """Get the configuratio of computer for user as a dictionary."""
-    from aiida.orm.backend import construct_backend
-
-    backend = construct_backend()
-    user = user or backend.users.get_automatic_user()
-
-    config = {}
-    try:
-        authinfo = backend.authinfos.get(computer, user)
-        config = authinfo.get_auth_params()
-    except NotExistent:
-        pass
-    return config
 
 
 class ComputerBuilder(object):
